// src/assessment/project/interface/project.interface.ts
export interface IProject {
  title: string;
  code: string;
  shortcode: string;
  country: string;
  state: string;
  city: string;
  department: string;
  telecommuting: boolean;
  url: string;
  application_url: string;
  shortlink: string;
  created_at: string;
  published_on: string;
  employment_type: string;
  industry: string;
  function: string;
  experience: string;
  education: string;
  repo: string;
}